# api-laravel-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IFRI-UAC notation, API Documentation",
        "description": "IFRI-UAC notation, API Documentation",
        "contact": {
            "email": "contact@equipe-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.43.5:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/formulaires-notations": {
            "get": {
                "tags": [
                    "FormulaireNotation"
                ],
                "summary": "FormulaireNotation - List",
                "description": "Returns list FormulaireNotation",
                "operationId": "formulaireNotationList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FormulaireNotationModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/formulaires-notations/create": {
            "post": {
                "tags": [
                    "FormulaireNotation"
                ],
                "summary": "FormulaireNotation - Create",
                "description": "Returns FormulaireNotation created",
                "operationId": "createFormulaireNotation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormulaireNotationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormulaireNotationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/formulaires-notations/{id}/update": {
            "put": {
                "tags": [
                    "FormulaireNotation"
                ],
                "summary": "FormulaireNotation - Update",
                "description": "Returns updated FormulaireNotation",
                "operationId": "listNotation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormulaireNotationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormulaireNotationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/formulaires-notations/{id}/delete": {
            "delete": {
                "tags": [
                    "FormulaireNotation"
                ],
                "summary": "FormulaireNotation - Delete",
                "description": "Returns null FormulaireNotation",
                "operationId": "deleteFormulaireNotation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "FormulaireNotation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormulaireNotationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/niveaux": {
            "get": {
                "tags": [
                    "Niveau"
                ],
                "summary": "Niveaux - List",
                "description": "Returns list niveau",
                "operationId": "niveauList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NiveauModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/niveaux/create": {
            "post": {
                "tags": [
                    "Niveau"
                ],
                "summary": "Niveaux - Create",
                "description": "Returns niveau created",
                "operationId": "createNiveau",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NiveauRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NiveauResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/niveaux/{id}/update": {
            "put": {
                "tags": [
                    "Niveau"
                ],
                "summary": "Niveaux - Update",
                "description": "Returns updated niveau",
                "operationId": "listNiveaux",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Niveau id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NiveauRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NiveauResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/niveaux/{id}/delete": {
            "delete": {
                "tags": [
                    "Niveau"
                ],
                "summary": "Niveaux - Delete",
                "description": "Returns null niveau",
                "operationId": "deleteNiveaux",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Niveau id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NiveauResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/formulaires-notations/{codeFormulaireNotation}/notations": {
            "get": {
                "tags": [
                    "Notation"
                ],
                "summary": "Notation - List",
                "description": "Returns list Notation",
                "operationId": "notationList",
                "parameters": [
                    {
                        "name": "codeFormulaireNotation",
                        "in": "path",
                        "description": "Code formulaire notation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListNotationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/formulaires-notations/{codeFormulaireNotation}/notations/{notationId}": {
            "get": {
                "tags": [
                    "Notation"
                ],
                "summary": "Notation - Item",
                "description": "Returns Notation",
                "operationId": "notationItem",
                "parameters": [
                    {
                        "name": "codeFormulaireNotation",
                        "in": "path",
                        "description": "Code formulaire notation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "notationId",
                        "in": "path",
                        "description": "Notation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FormulaireNotationModel": {
                "title": "FormulaireNotation Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "code": {
                        "title": "code",
                        "description": "code of Notation",
                        "type": "string",
                        "example": "klmd-huib-sqwg-sgvx-kddu"
                    },
                    "classe": {
                        "title": "classe",
                        "description": "classe of Notation"
                    },
                    "matiere": {
                        "title": "matiere",
                        "description": "classe of Notation"
                    },
                    "enseignant": {
                        "title": "enseignant",
                        "description": "enseignant of Notation"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "FormulaireNotation"
                }
            },
            "FormulaireNotationRequest": {
                "title": "Save FormulaireNotation Request",
                "required": [
                    "classe_id",
                    "matiere_id",
                    "enseignant_id"
                ],
                "properties": {
                    "classe_id": {
                        "title": "classe_id",
                        "description": "classe_id of the Notation",
                        "type": "integer",
                        "example": "19"
                    },
                    "matiere_id": {
                        "title": "matiere_id",
                        "description": "matiere_id of the Notation",
                        "type": "integer",
                        "example": "6"
                    },
                    "enseignant_id": {
                        "title": "enseignant_id",
                        "description": "enseignant_id of the Notation",
                        "type": "integer",
                        "example": "4"
                    }
                },
                "type": "object"
            },
            "FormulaireNotationResponse": {
                "title": "FormulaireNotation Response",
                "properties": {
                    "formulaireNotation": {
                        "$ref": "#/components/schemas/FormulaireNotationModel"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "FormulaireNotationResponse"
                }
            },
...
            "NiveauModel": {
                "title": "Niveau Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "libelle": {
                        "title": "libelle",
                        "description": "libelle of niveau",
                        "type": "string",
                        "example": "L3"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Niveau"
                }
            },
            "NiveauRequest": {
                "title": "Save Niveau Request",
                "required": [
                    "libelle"
                ],
                "properties": {
                    "libelle": {
                        "title": "libelle",
                        "description": "Libelle of the niveau",
                        "type": "string",
                        "example": "M3"
                    }
                },
                "type": "object"
            },
            "NiveauResponse": {
                "title": "Niveau Response",
                "properties": {
                    "niveau": {
                        "title": "niveau"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NiveauResponse"
                }
            },
            "LigneNotationModel": {
                "title": "LigneNotation Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "critere": {
                        "title": "critere"
                    },
                    "reponse": {
                        "title": "reponse",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LigneNotation"
                }
            },
            "ListNotationResponse": {
                "title": "List notation Response",
                "properties": {
                    "notations": {
                        "title": "niveau",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NotationModel"
                        }
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ListNotationResponse"
                }
            },
            "NotationModel": {
                "title": "Notation Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "lignes": {
                        "title": "lignes",
                        "description": "lignes of Notation",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LigneNotationModel"
                        }
                    },
                    "formulaire_notation": {
                        "title": "lignes",
                        "description": "lignes of Notation"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Notation"
                }
            },
            "NotationResponse": {
                "title": "Notation Response",
                "properties": {
                    "notation": {
                        "$ref": "#/components/schemas/NotationModel"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NotationResponse"
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "The value of access_token received from login: Ex: Bearer access_token_value",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}